Changed heightmap algorithm
So the heightmap algorithm has been changed.
The idea is that the level gets generated as a set of squares, each
square being in the end a face on the final terrain.

Actually, I'm a bit tired, and I might scrap this algorithm anyways, so
more details when it actually WORKS! :)