Sensor-controlled hands and sword!
So, after many a battle with the sdk and compiler, I have managed it!
Well, kinda.

Hands captured by the sensor follow your character, and stay (mostly)
within your field of view.

I've also managed to make a sword that (kinda) stays in your hand. It
works fine when everything is on default, but when I increase the
transform sensitivity for the tracking (i. e. moving my hand by 1 cm
moves the virtual one by 4, or 8...), the sword follows a point a
distance away from the hand...

Will see what i can do about that later.