Improved "Immortal Object" message and collider stuff
So... Too tired to write details, as always.

Made it so the "Immortal Object" signs appear only once every 1.4
seconds (to avoid them spawning over each-other)
Fixed the partly-transparent material for the signs (you could see the
edges of the rectangle)

Made the vase a rigidbody that can be knocked over with the sword /
player. if hit by sword, it will be removed 4 seconds after.

Also messed about a lot with collision physics, and made other
adjustments here and there which i probably won't ever remember...