Redundant Health and Damage Script, Floating Health UI and Smooth Health Decrease
A number of changes today.

I've added a UI system inspired by that of SAO. (Currently only applied
to pots :) )

Given that I wanted the UI to be curved and have custom sprites, I ended
up making 101 separate sprites in Inkspace.

Adding this made me make a redundant Health System that can be easily
applied to new props / enemies.

Each UI object gets parented to the GameObject it will monitor by the
PropPlacer.cs script. In order for the UI to work, the GameObject needs
to have an HealthController.cs script where you simply define the
objects starting health points, and call a TakeDamage() function from
other scripts to damage the object.
The HealthMonitor.cs script in the UI object will do the rest by
constantly checking the Hit Points of its parent and updating
accordingly.

There is a total of 101 sprites for the health display (from 0 to 100%),
with the percentage automatically calculated.
part of the script makes sure health decreases (no support for increases
yet) are displayed smoothly, with the Health Bar going down at a speed
proportional to the damage taken (there is a parameter to change the
overall speed of the smooth decrease effect).

I know this is a bit overkill and probably horribly inefficient, but I
have no better ideas ><

On a side note, the player's rigidbody no longer spazzes out when you
walk over a pot because I finally got around to increasing its mass by
40 :)