Buttons Can Do Things, UI Relative to Head and Single-Press Buttons
The UIMenuHolder is now a child of the main camera, and therefore will
follow the player's view, as it should.

Modified the button scripts from the Leap SDK so that you can now check
whether a button is active from other scripts (ButtonDemoToggle.isActive
or ButtonDemoNoToggle.isActive).
Also copied the ButtonBaseToggle and the ButtonDemoToggle scripts into
NoToggle versions in which the button will not stay on after released,
but only while pressed (can't beleaive they did not have that in their
example widgets!).

Oh, and I started working on nested menus. Visual - only UI can already
be nested simply by parenting it to the OnGraphics child object of a
button, but buttons will need to be tweaked, since for now both on and
off states get activated when a nested button appears.