OVR Camera Rig, Head-Mounted Leap Traker, UI Element Prefabs
So I've converted to the OVR Camera rig instead of the standard camera,
and made sure the hand tracking will work in VR by switching to a
HeadMountedController prefab. Switching back to regular camera &
tracking just requires to activate some objects in the FPPlayer object.

In other news, to make UI building easier, I made some prefabs. There is
a MenuBackground prefab for the visual UI (unfortunately, it seems like
you will have to scale the text background manually every time), and a
MenuText prefab for, you guessed it, the text in sub-menus.