Corrupted Cubes, Structures, and The Flat Camera Return
A number of small optimisations here and there.

Generation chances are now stored in a SpawnConfig on each object in the
form 1/probability rather than in the map generator script. They are
overrideable.

Reverted to regular non-rift camera since it makes things compile
faster.

Added Structure tag (or maybe in the previous commit). Doesn't do stuff
quite _yet_.

Minimap has returned, so I added a special MInimap layer to the
rendering, which is invisible to the main camera.
There is a prism/triangle on that layer that follows the player and
turns with them.

Background code development for substructure generation including
higher-resolution generator maps for structures. Resolution could
eventually be customizable.

Script to eventually allow keyboard menu interaction. (Nothing works
yet).

Glitched/corrupt cube mesh added for eventual structure tests.

Skybox/lighting change to look impressive. :)

A lot of kill streaks in Ghost in the Shell!