Pathfinding Beginnings
WIP (Nothing functional yet!)

Added a new script PathGridGenerator. It will eventually be used to
generate a Undirected Graph to be used in pathfinding.

The principle (for now) is as follows. The grid is generated with the
rest of the world. Each Node has a Coords attribute (x,y,z) and a way of
telling which nodes it is linked to, and possibly an associated cost
(terrain that slows down, slopes).

As of right now, the Graph class keeps track of all of its nodes, which
in turn keep track of each of their own coordinates. A better
implementation would be to use a Map/Dictionary to map coordinates
(preferrably approximate ones, using doubles) to Nodes, which would
allow AIs to access a Node using their coordinates only with a quick
lookup, and will make it much less computationally intensive to keep
track of all the nodes and avoid doubles (Ordered Dictionaty would be
the best choice for now I think).

P.S.: Corrected random typo somewhere I think.