Pathfinding Grid Alpha
And I'm baaaack!
Here is a preliminary implementation of the Pathfinding Grid using
Sorted DIctionaries. It's not connected to anything yet, so I haven't
tested it, but here is the gist of it.
Everything is stored as a SortedDictionary<Coords,Node>, using a custom
comparison method that does approximate comparison between coordinates,
so 2 nearby coordinates are counted as identical.
Each node stores an associated coordinate (maybe unnecessary) and a set
of neigbours that are connected to it directly, stored as a
HashSet<Coords> with a custom compare method as well.
SortedDictonary seems to  be a good choice, since it has O{log n) lookup
efficiency, and those coordinate grids may eventually grow large....